Refactor public templates into modules
